Fix Ticker types

+++ src/main/java/cc/bitbank/entity/Ticker.java
@@ -10,11 +10,11 @@ import java.util.Date;
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Ticker extends Data {
-    public int sell;
-    public int buy;
-    public int high;
-    public int low;
-    public int last;
+    public BigDecimal sell;
+    public BigDecimal buy;
+    public BigDecimal high;
+    public BigDecimal low;
+    public BigDecimal last;
     public BigDecimal vol;
     public Date timestamp;
